Fixes img issues in ViewMenuItem component from recent merge
