+++ src/components/MenuComponents/ViewMenuItem/ViewMenuItem.module.css
@@ -124,8 +124,8 @@
 }
 
 .title img {
-  width: 375px;
-  height: 375px;
+  width: 100%;
+  height: 100%;
 }
 
 @media (max-width: 500px) {
@@ -137,7 +137,7 @@
 
 @media (min-width: 768px) and (max-width: 1023px) {
   .title img {
-    width: 284px;
-    height: 284px;
+    width: 100%;
+    height: 100%;
   }
 }
